Remove GY21.hpp

+++ automation_code_builder/drivers/arduino/LED.hpp
@@ -7,7 +7,7 @@ class LED : public virtual IOutput
    
     private:
         long pin;
-        long value{0};
+        int value{0};
     public:
         LED(const char* id, const DeviceSettings& settings): IDevice(id, "LED", settings), pin(settings.get_config().at_or_default("pin", -1)),
         value(settings.get_config().at_or_default("init_value", 0)){}
